% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ELISA.R
\name{clean_ELISA}
\alias{clean_ELISA}
\title{Cleans ELISA PLate Plan}
\usage{
clean_ELISA(path)
}
\arguments{
\item{path}{input file directory}
}
\value{
A tibble with tidy dataset
}
\description{
Read in individuals ELISA Plate plans and performs data set merge.
}
\examples{
clean_ELISA(path = "(C:/Users/hkiprono/3.0 CHMI/)")

}
