% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_Data.R
\name{cod2dig}
\alias{cod2dig}
\title{Encerar codigo PROV-CANT-PARR}
\usage{
cod2dig(vector)
}
\arguments{
\item{vector}{vector de datos}
}
\description{
Funcion encera el codigo provincia, canton, parroquia.
}
