% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_Data.R
\name{constante}
\alias{constante}
\title{Analisis de variables constantes}
\usage{
constante(x)
}
\arguments{
\item{x}{vector de datos}

\item{corr.max}{numero entre 0 y 1}
}
\description{
Funcion que analiza el porcentaje de variables constantes en una lista o data.frame.
}
