% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_Data.R
\name{DVarCorr}
\alias{DVarCorr}
\title{Correlacion (Correlacion superior a un valor dado)}
\usage{
DVarCorr(data, corr.max = 0.75)
}
\arguments{
\item{data}{list, data.frame o data.table}

\item{corr.max}{numero entre 0 y 1}
}
\value{
list, data.frame o data.table
}
\description{
Funcion que indica fuerza y direccion de una relacion lineal y proporcionalidad entre dos variables.
}
