% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_Files.R
\name{Import}
\alias{Import}
\title{Import Excel files}
\usage{
Import(path, pattern = c(".xls", ".xlsx", ".csv"))
}
\arguments{
\item{path}{input file directory}

\item{pattern}{input pattern as either .xls, .xlsx or csv}
}
\value{
A tibble with tidy dataset
}
\description{
Read multiple excel data sets to R
}
\examples{
clean_ELISA(path = "C:/Users/hkiprono/Desktop/R_packages/test_data_TidyElisa",pattern = c(".xls",".xlsx"))


}
