% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_Data.R
\name{num_anios}
\alias{num_anios}
\title{Diferencia de fechas en anios}
\usage{
num_anios(end_date, start_date)
}
\arguments{
\item{end_date}{fecha final}

\item{start_date}{fecha inicial}
}
\description{
Funcion que realiza la diferencia de fechas en anios.
}
