% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_Data.R
\name{num_meses}
\alias{num_meses}
\title{Diferencia de fechas en meses}
\usage{
num_meses(end_date, start_date)
}
\arguments{
\item{end_date}{fecha final}

\item{start_date}{fecha inicial}
}
\description{
Funcion que realiza la diferencia de fechas en meses.
}
