% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_Data.R
\name{reemplazo_col}
\alias{reemplazo_col}
\title{Funcion de reemplazo NA's}
\usage{
reemplazo_col(dt, vars, valor)
}
\arguments{
\item{dt}{list, data.frame, data.table}

\item{vars}{variables de la base de datos}

\item{valor}{numero por el cual se reemplaza los NA's}
}
\description{
Funcion que reemplaza NA's de una base de datos con un valor indicado.
}
