% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_Data.R
\name{TestKS}
\alias{TestKS}
\title{KS (Coeficiente de Kolmogorov-Smirvov)}
\usage{
TestKS(x, y)
}
\arguments{
\item{x}{variable continua}

\item{y}{variable continua}
}
\description{
Funcion para seleccionar variables que generan mayor divergencia entre las distribuciones de dos muestras aleatorias.
}
