% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidy_Data.R
\name{TestVI}
\alias{TestVI}
\title{IV (Valor de Informacion)}
\usage{
TestVI(x, y)
}
\arguments{
\item{x}{variable categorica}

\item{y}{variable categorica}
}
\description{
Funcion que mide el poder predectivo de una variable categorica para  problemas de clasifciacion binaria.
}
