% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyname.R
\name{tidyname}
\alias{tidyname}
\title{tidy data for OLS and so on.}
\usage{
tidyname(dta)
}
\arguments{
\item{dta}{is data set(data.table,data.frame,tibble)}
}
\value{
explain what is returned by this function.
}
\description{
tidy data for OLS and so on.
}
